requirements.txt
langchain
openai
pyaudio
python-dotenv
RealtimeTTS
RealtimeSTT
